splitting libs
