+++ lib/tf/teraterm.keyboard.cnf
